Locust/requirements.txt
autopep8==1.6.0
black==22.10.0
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
ConfigArgParse==1.5.3
Flask==2.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==2.0.2
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.0.3
locust==2.12.0
lorem==0.1.1
MarkupSafe==2.1.1
msgpack==1.0.3
mypy-extensions==0.4.3
paho-mqtt==1.6.1
pathspec==0.10.1
platformdirs==2.5.3
psutil==5.9.0
pycodestyle==2.8.0
pyzmq==22.3.0
requests==2.27.1
roundrobin==0.0.2
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zope.event==4.5.0
zope.interface==5.4.0
